+++ src/main/java/com/jk18_7/sim/login/service/LoginService.java
@@ -24,7 +24,7 @@ public interface LoginService {
 
     void updateUser(Users users);
 
-    Users getUser(long uid);
+    Users getUser(int uid);
 
     List<Users> getList();
