+++ src/main/java/com/jk18_7/sim/login/service/impl/LoginServiceImpl.java
@@ -48,7 +48,7 @@ public class LoginServiceImpl implements LoginService {
     }
 
     @Override
-    public Users getUser(long uid) {
+    public Users getUser(int uid) {
         return loginDao.getUser(uid);
     }
